Name change and simplification to remove 'state' helpers
Makes this mixin useful only for the general case of dealing with
observable sequences. Removes any references to state and filtering
objects for key changes, etc, which can be achieved with generic
functional libs.
